Rename `pullRequest` to `pullNumber`

+++ src/steps/maybeSetupRepo.ts
@@ -20,12 +20,12 @@ export async function maybeSetupRepo({
   repoName: string;
   username: string;
 }) {
-  const spinner = ora('').start();
   const isAlreadyCloned = await repoExists({ owner, repoName });
 
-  try {
-    // clone repo if folder does not already exists
-    if (!isAlreadyCloned) {
+  // clone repo if folder does not already exists
+  if (!isAlreadyCloned) {
+    const spinner = ora().start();
+    try {
       const spinnerCloneText = 'Cloning repository (one-time operation)';
       spinner.text = `0% ${spinnerCloneText}`;
       await mkdirp(getRepoOwnerPath(owner));
@@ -38,21 +38,20 @@ export async function maybeSetupRepo({
           spinner.text = `${progress}% ${spinnerCloneText}`;
         }
       });
-      spinner.text = `100% ${spinnerCloneText}`;
-    } else {
-      spinner.text = `Setup remotes for repository`;
+      spinner.succeed(`100% ${spinnerCloneText}`);
+    } catch (e) {
+      spinner.fail();
+      await deleteRepo({ owner, repoName });
+      throw e;
     }
+  }
 
-    // ensure remote are setup with latest accessToken
+  // ensure remote are setup with latest accessToken
+  await deleteRemote({ owner, repoName, username });
+  await addRemote({ owner, repoName, username, accessToken });
+
+  if (username !== owner) {
     await deleteRemote({ owner, repoName, username: owner });
-    await deleteRemote({ owner, repoName, username });
     await addRemote({ owner, repoName, username: owner, accessToken });
-    await addRemote({ owner, repoName, username, accessToken });
-
-    spinner.succeed();
-  } catch (e) {
-    spinner.fail();
-    await deleteRepo({ owner, repoName });
-    throw e;
   }
 }
